requirements.txt
Flask==3.0.3
Flask-PyMongo==2.3.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.1
dnspython==2.7.0
pymongo==4.4.0
Flask-Cors==4.0.0